Extract exec, avoid circular dependency
